requirements.txt
streamlit>=1.20
google-generativeai>=0.5.0
faiss-cpu>=1.7.3
numpy
